requirements.txt
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
sentence-transformers>=2.7.0
requests>=2.31.0
numpy>=1.26.0
openai>=1.14.2
pytest>=7.4.0
fastapi>=0.110.0
uvicorn>=0.24.0
python-multipart>=0.0.6
celery>=5.3.6
psycopg[binary]>=3.1.18
redis>=5.0.4
websockets>=11.0.3
pydantic>=2.6.1
